Fix Docker permission issue - remove non-root user
The container was failing with 'Permission denied' when trying to execute gunicorn.
This was because gunicorn was installed to /root/.local/bin but the container
was switching to a non-root user that couldn't access it.

Cloud Run provides container isolation, so running as root is acceptable.
Removed the non-root user creation to fix the permission issue.